Fix-reg-order (#357)
* feat: Update order_id column in RegistrationOrder model

Refactor the `order_id` column in the `RegistrationOrder` model to use the `Computed` function from SQLAlchemy. This change improves the efficiency and readability of the code by calculating the `order_id` value based on the `event_id`, `attendee_id`, and `id` columns.

* feat: Refactor order_id column in RegistrationOrder model

Refactor the `order_id` column in the `RegistrationOrder` model to use the `Computed` function from SQLAlchemy. This change improves the efficiency and readability of the code by calculating the `order_id` value based on the `event_id`, `attendee_id`, and `id` columns.

* Refactor order_id column in RegistrationOrder model